changed the setLong to setInt for Oracle bug
git-svn-id: svn+ssh://neptune.elsevier.com/usr/local/subversion/repos/baja/trunk@6341 aa8a525f-4958-0410-8862-b993faadc08d

+++ eijava/eilib/src/org/ei/domain/PageCache.java
@@ -28,7 +28,7 @@ public class PageCache
 {
 
     private String sessionID;
-    private static long timeHolder = 1L;
+    private static int timeHolder = 1;
 
     /**
     *   Constructor which takes session id as argument.
@@ -195,8 +195,8 @@ public class PageCache
             proc = con.prepareCall("{ call PageCache_writePage(?,?,?,?,?)}");
             proc.setString(1,sessionID);
             proc.setString(2,searchID);
-            proc.setLong(3,pageNo);
-            proc.setLong(4,timeHolder);
+            proc.setInt(3,pageNo);
+            proc.setInt(4,timeHolder);
             proc.setString(5,lString);
             proc.executeUpdate();
         }
